refactor: 🔧 update project structure documentation and rename use-cases to usecases
