+++ scripts/generate-structure-docs.sh
@@ -11,7 +11,10 @@ doc_file="docs/src/content/docs/structure.md"
 date_only="$(date -u '+%Y-%m-%d')"
 
 cat > "$doc_file" <<EOF
-# Project Structure
+---
+title: Project Structure
+description: Overview of the project structure and organization.
+---
 
 <!-- markdownlint-disable-next-line MD036 -->
 **Auto-generated by scripts/generate-structure-docs.sh on ${date_only}**
